Created AdminTeacherCommonsController instead and moved some methods from AdminController and TeacherController there.
