+++ src/main/java/com/kasiapetka/topicsmanager/controllers/AdminController.java
@@ -171,10 +171,7 @@ public class AdminController {
         return teacherService.listActiveTeachers();
     }
 
-    @GetMapping("/api/admin/students")
-    List<Student> listStudents() {
-        return studentService.listActiveStudents();
-    }
+
 
     @PutMapping("/api/admin/deleteTeacher")
     ResponseEntity<?> deleteTeacher(@Valid @RequestBody Long id) {
