+++ src/main/java/com/kasiapetka/topicsmanager/model/Subject.java
@@ -17,7 +17,6 @@ public class Subject {
     private long id;
     @NotNull
     private String name;
-
     private String summary;
 
     @JsonManagedReference
